melissa-engine/pyproject.toml
[project]
name = "melissa-engine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "PyNaCl>=1.5",
  "pyyaml>=6.0",
  "jsonschema>=4.23",
]

[project.scripts]
melissa = "src.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
